requirements.txt
# Database drivers
mysql-connector-python==8.2.0
psycopg[binary]>=3.1.0

# Database abstraction layer
SQLAlchemy>=2.0.25

# Environment Management
python-dotenv==1.0.0

# Progress & Monitoring
tqdm==4.66.1
colorama==0.4.6

# Utilities
tabulate==0.9.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Optional system info
psutil>=5.9.0
